+++ test/lib/global_uid_test.rb
@@ -58,7 +58,7 @@ describe GlobalUid do
     describe "without explicit parameters" do
       describe "with global-uid enabled" do
         before do
-          GlobalUid::Base.global_uid_options[:storage_engine] = "InnoDB"
+          GlobalUid.configuration.storage_engine = "InnoDB"
           CreateWithNoParams.up
           @create_table = show_create_sql(WithGlobalUID, "with_global_uids").split("\n")
         end
@@ -113,7 +113,7 @@ describe GlobalUid do
 
       describe "with global-uid disabled, globally" do
         before do
-          GlobalUid::Base.global_uid_options[:disabled] = true
+          GlobalUid.configuration.disabled = true
           CreateWithNoParams.up
         end
 
@@ -202,7 +202,7 @@ describe GlobalUid do
 
   describe "With InnoDB engine" do
     before do
-      GlobalUid::Base.global_uid_options[:storage_engine] = "InnoDB"
+      GlobalUid.configuration.storage_engine = "InnoDB"
       CreateWithNoParams.up
       CreateWithoutGlobalUIDs.up
     end
@@ -219,8 +219,9 @@ describe GlobalUid do
       CreateWithoutGlobalUIDs.up
 
       @notifications = []
-      GlobalUid::Base.global_uid_options[:notifier] = Proc.new do |exception, message|
-        GlobalUid::Base::GLOBAL_UID_DEFAULTS[:notifier].call(exception, message)
+      original = GlobalUid.configuration.notifier
+      GlobalUid.configuration.notifier = Proc.new do |exception, message|
+        original.call(exception, message)
         @notifications << exception.class
       end
     end
@@ -240,7 +241,7 @@ describe GlobalUid do
 
     describe "with increment exceptions suppressed " do
       before do
-        GlobalUid::Base.global_uid_options[:suppress_increment_exceptions] = true
+        GlobalUid.configuration.suppress_increment_exceptions = true
       end
 
       it "allows the increment to be updated" do
@@ -296,15 +297,16 @@ describe GlobalUid do
       describe 'when the auto_increment_increment changes' do
         before do
           @notifications = []
-          GlobalUid::Base.global_uid_options[:notifier] = Proc.new do |exception, message|
-            GlobalUid::Base::GLOBAL_UID_DEFAULTS[:notifier].call(exception, message)
+          original = GlobalUid.configuration.notifier
+          GlobalUid.configuration.notifier = Proc.new do |exception, message|
+            original.call(exception, message)
             @notifications << exception.class
           end
         end
 
         describe "and all servers report a value other than what's configured" do
           it "raises an exception when configuration incorrect during initialization" do
-            GlobalUid::Base.global_uid_options[:increment_by] = 42
+            GlobalUid.configuration.increment_by = 42
             GlobalUid::Base.disconnect!
             assert_raises(GlobalUid::NoServersAvailableException) { test_unique_ids(10) }
             assert_includes(@notifications, GlobalUid::InvalidIncrementException)
@@ -332,7 +334,7 @@ describe GlobalUid do
 
           it "doesn't cater for increment_by being increased by a factor of x" do
             GlobalUid::Base.with_servers do |server|
-              server.connection.execute("SET SESSION auto_increment_increment = #{GlobalUid::Base::GLOBAL_UID_DEFAULTS[:increment_by] * 2}")
+              server.connection.execute("SET SESSION auto_increment_increment = #{GlobalUid.configuration.increment_by * 2}")
             end
             # Due to multiple processes and threads sharing the same alloc server, identifiers may be provisioned
             # before the current thread receives its next one. We rely on the gap being divisible by the configured increment
@@ -463,7 +465,7 @@ describe GlobalUid do
 
     describe "without shuffling connections" do
       before do
-        GlobalUid::Base.global_uid_options[:connection_shuffling] = false
+        GlobalUid.configuration.connection_shuffling = false
       end
 
       it "increment sequentially" do
@@ -488,14 +490,14 @@ describe GlobalUid do
       end
 
       after do
-        GlobalUid::Base.global_uid_options[:connection_shuffling] = true
+        GlobalUid.configuration.connection_shuffling = true
       end
     end
 
 
     describe "with connection shuffling" do
       before do
-        GlobalUid::Base.global_uid_options[:connection_shuffling] = true
+        GlobalUid.configuration.connection_shuffling = true
       end
 
       it "increment sequentially" do